# schemas/openapi.json
{
  "components": {
...
    "responses": {
      "Error400": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "`ErrorResponse` with `code` in: `invalid_request`"
      },
      "Error401": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "`ErrorResponse` with `code` in: `unauthenticated`"
      },
      "Error403": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "`ErrorResponse` with `code` in: `forbidden`"
      },
      "Error404": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "`ErrorResponse` with `code` in: `not_found`"
      },
      "Error409": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "`ErrorResponse` with `code` in: `version_conflict`, `task_not_retryable`, `status_conflict`, `gate_not_passed`"
      },
      "Error422": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "`ErrorResponse` with `code` in: `schema_violation`, `idempotency_payload_mismatch`"
      },
...
    },
    "schemas": {
...
      "DocumentListResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "minimum": 0,
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "DocumentListResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "OIDC-compatible token (ADR-0001, INV-AUTH-01): the server verifies issuer, audience, expiry and signature; department and scopes are mapped server-side from `sub`, never taken from the request.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Public REST contract of MedOps Copilot (baseline 5.7). Identity always comes from the verified bearer token; request bodies never carry `dept` or `scopes`. Every error uses `ErrorResponse` (code, message, trace_id, retryable) with no internal detail. Final medical answers are never streamed before the safety check completes (INV-SAF-05). Admin document endpoints (M1), trace replay (M3-08) and policy approval (M4) are added to this document when their models exist.",
    "title": "MedOps Copilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/documents": {
      "get": {
        "description": "Metadata, status, effective window and read departments of every document, optionally filtered by `dept`, `status` or `family_id`; never returns chunk content. Requires the admin role (DEC-012).",
        "operationId": "listDocuments",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            },
            "description": "Document list."
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          }
        },
        "summary": "List documents (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/documents/{doc_id}": {
      "get": {
        "description": "Metadata plus the read ACL and the 20 most recent audit rows. Requires the admin role.",
        "operationId": "getDocument",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description": "Document detail."
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          }
        },
        "summary": "Document detail with ACL and audit tail (admin)",
        "tags": [
          "admin"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/DocumentIdPath"
        }
      ]
    },
    "/admin/documents/{doc_id}/acl": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DocumentIdPath"
        }
      ],
      "patch": {
        "description": "The owner department keeps read access; withdrawn documents are read-only. Each change writes an audit row and an outbox event that bumps the cache epoch of every department involved. `Idempotency-Key` (baseline 3.2). Scope is authenticated principal + route + key; the server stores the canonical request hash. Same key and same payload returns the original receipt with the original status code, also while the first request is still processing (never 409). Same key with a different payload returns 422 `idempotency_payload_mismatch`. The record and the idempotency entry are created in one transaction. Retention is configured by IDEMPOTENCY_KEY_TTL_SECONDS (default 604800 seconds, never below 86400 seconds) and never shorter than the task lifetime.",
        "operationId": "changeDocumentAcl",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentAclRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentAclResponse"
                }
              }
            },
            "description": "Departments granted, revoked and the resulting read ACL."
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          }
        },
        "summary": "Grant or revoke department read access (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/documents/{doc_id}/status": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DocumentIdPath"
        }
      ],
      "patch": {
        "description": "`activate`: draft to active at `effective_date` (the family must have no other active version); `archive`: active to archived, closing the window at `effective_date`; `withdraw`: draft to the terminal withdrawn state. Illegal transitions return `409 status_conflict`. The reason is written to the audit row and an outbox event invalidates the retrieval cache. Uploads are not an HTTP operation (the ingestion pipeline stays offline, DEC-012). `Idempotency-Key` (baseline 3.2). Scope is authenticated principal + route + key; the server stores the canonical request hash. Same key and same payload returns the original receipt with the original status code, also while the first request is still processing (never 409). Same key with a different payload returns 422 `idempotency_payload_mismatch`. The record and the idempotency entry are created in one transaction. Retention is configured by IDEMPOTENCY_KEY_TTL_SECONDS (default 604800 seconds, never below 86400 seconds) and never shorter than the task lifetime.",
        "operationId": "changeDocumentStatus",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description": "Document detail after the change."
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          }
        },
        "summary": "Activate, archive or withdraw a document (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/policies/candidates": {
      "get": {
        "description": "Candidate diffs (prompt, rule, skill or retrieval parameters) with their evidence. Requires the approver or admin role.",
        "operationId": "listPolicyCandidates",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyListResponse"
                }
              }
            },
            "description": "Candidates."
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          }
        },
        "summary": "List policy candidates awaiting a decision (approver / admin)",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}